pyproject.toml
[tool.poetry]
name = "global-connected-tide-gauge-station-clustering"
version = "0.1.0"
description = ""
authors = ["Johanna Hillebrand <johanna.hillebrand@hhu.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scipy = "^1.13.0"
loguru = "^0.7.2"
tqdm = "^4.66.4"
geopandas = "^0.14.4"
matplotlib = "^3.8.4"
shapely = "^2.0.4"
networkx = "^3.3"
searoute = "^1.3.1"
netcdf4 = "^1.6.5"
astropy = "^6.1.0"
jupyter = "^1.0.0"
ccrs = "^1.0.0"
cartopy = "^0.23.0"
rioxarray = "^0.15.5"
haversine = "^2.8.1"
seaborn = "^0.13.2"
regionmask = "^0.12.1"
srai = "^0.7.5"
plotly = "^5.22.0"
scikit-learn = "^1.5.1"
dask = "^2024.7.1"
geopy = "^2.4.1"
xeofs = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
